Fix compatibility with Test-Fatal 0.016
Test-Fatal 0.016 sets $Carp::MaxArgNums to -1, which makes function
arguments invisible in the stack trace and this in turn causes t/debug.t
to fail. To fix this, explicitly set $Carp::MaxArgNums to 2 in the test
function to restore visibility of function arguments. This change allows
the test to work with Test-Fatal 0.016 and also older versions.